generate_fake_login/requirements.txt
PyJWT===1.7.1
boto3===1.12.20
pyjwt-rsa===0.3.0